Equality breaker moved to GUI directory
- Moved EqualityBreaker to the GUI directory
- Now it is based unto the Auxiliary_Window class for better DRY code.
