+++ auxiclean/gui/__init__.py
@@ -1,0 +1,2 @@
+from .auxiliary_window import AuxiliaryWindow
+from .equality_breaker import EqualityBreaker
